Add 404 page, re-arrange layouts
