+++ gatsby-config.js
@@ -51,7 +51,7 @@ module.exports = {
       resolve: 'gatsby-mdx',
       options: {
         defaultLayouts: {
-          default: require.resolve('./src/layouts/Article.js'),
+          default: require.resolve('./src/templates/Article.js'),
         },
         gatsbyRemarkPlugins: [
           {
