+++ src/assets/styles/site.scss
@@ -9,3 +9,4 @@
 //
 
 @import './modules/docs';
+@import './modules/not-found';
