+++ src/pages/inline.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import FoundationLayout from '../layouts/Foundation';
+import BaseLayout from '../layouts/BaseLayout';
 import { Alert, Button, Inline } from '../components';
 import Live from '../docs/Live';
 
@@ -26,7 +26,7 @@ const InlineScope = {
 };
 
 const InlineDocs = () => (
-  <FoundationLayout pageTitle="Inline">
+  <BaseLayout pageTitle="Inline">
     <p>Sometimes your content doesn&apos;t belong in a grid system. So, for when you have more &quot;free-form&quot; content that you&apos;d like proper spacing horizontally and vertically when they stack, we created the Inline component. You can modify alignment by using our <a href="/flex/">Flex</a> utilities.</p>
 
     <Live
@@ -39,7 +39,7 @@ const InlineDocs = () => (
     <Alert type="warning" title="Note‍">
       <p>There should be no whitespace modifiers attached to the Inline component or its direct children. You can attach them on adjacent DOM or by wrapping the component in a <code>&lt;div&gt;</code></p>
     </Alert>
-  </FoundationLayout>
+  </BaseLayout>
 );
 
 export default InlineDocs;
