+++ src/templates/Article.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import FoundationLayout from './Foundation';
+import BaseLayout from '../layouts/BaseLayout';
 
 const Article = (props) => {
   const { children } = props;
 
   return (
-    <FoundationLayout>
+    <BaseLayout>
       {children}
-    </FoundationLayout>
+    </BaseLayout>
   );
 };
